Add bsTab to restaurant edit view

+++ resources/js/components/images/ListImagesComponent.vue
@@ -1,14 +1,18 @@
 <template>
   <div class="row">
-    <div v-for="image in images" :key="image.id" class="col-md-2 my-2">
-      <img :src="image.url" class="img-fluid rounded" width="200" />
-      <button
-        type="button"
-        class="btn btn-outline-danger btn-sm mt-2"
-        @click="deleteImage(image.id)"
-      >
-        Delete
-      </button>
+    <div v-for="image in images" :key="image.id" class="col-md-2 my-2 mx-auto">
+      <div class="text-center">
+        <img :src="image.url" class="img-fluid rounded" width="200" />
+      </div>
+      <div class="text-center">
+        <button
+          type="button"
+          class="btn btn-outline-danger btn-sm mt-2"
+          @click="deleteImage(image.id)"
+        >
+          <i class="bi bi-trash"></i>
+        </button>
+      </div>
     </div>
   </div>
 </template>
@@ -29,6 +33,12 @@ export default {
 
   methods: {
     deleteImage(id) {
+      let confirmDelete = confirm(
+        "Tu es sûr de vouloir supprimer cette image ?"
+      );
+      if (!confirmDelete) {
+        return;
+      }
       this.loading = true;
       axios.delete(`/api/restaurants/images/${id}`).then(() => {
         this.loading = false;
